analytics: squash loss-of-precision warnings
And, in passing, avoid duplicating the logic to mangle the integer value
into the string label.

+++ src/endless/Analytics.h
@@ -10,7 +10,7 @@ public:
 
 	void sessionControl(BOOL start);
 	void screenTracking(const CString &name);
-	void eventTracking(const CString &category, const CString &action, const CString &label = CString(), int value = -1);
+	void eventTracking(const CString &category, const CString &action, const CString &label = CString(), LONGLONG value = -1);
 	void exceptionTracking(const CString &description, BOOL fatal);
 
 	void setLanguage(const CString &language);
